Nav bar working
Got the nav bar to work with switching between scenes. Persisting issue:
there is a special case for clicking on home. It is supposed to go back
to the home scene, which it does. Once it does this however, the nav bar
stops working again... Will work the issue out hopefully in next commit.